refactor(frontend): extract useStudioTabs composable (P2-2 F14)
Completes the F8 residue: the tab workspace lifecycle (~413 lines) moves
from DataStudioNew.vue into composables/useStudioTabs.js — createTabState,
openTableTab, loadTabData, hydrateRestoredTableTabs, disposeTabResources,
handleTabRemove/CloseLeft/CloseRight/CloseAll, handleTabAdd, and the
resolve/insert-index helpers. Shared refs (openTabs/activeTab/tabStates)
stay owned by the component and are passed by reference, matching the
existing composable wiring; all deps are declared before the assembly
point so the previous lazy forward references to openTableTab now close
the loop. provide('dataStudioCtx') keys unchanged.

New __tests__/useStudioTabs.spec.js covers tab state shape, add/remove/
close-left/right/all transitions, open-existing vs open-new table tabs,
loadTabData dedupe and failure handling, and restore hydration (13 tests).

DataStudioNew.vue: 2759 -> 2380 lines.

Verified: npm run lint (0 errors), npm run test (28 files / 134 passed,
incl. DataStudioNew mount smoke), npm run build; demo-mode Playwright
smoke (no real backend): catalog expand -> open table tab (right panel
renders), new/switch/close query tabs, reload restores tabs. Only console
error is the pre-existing /dataagent/widget bundle proxy 500.
